Tests do not need to sleep before sync status checks; Add sync option for delete_table; Add delete_table_synchronously; Handle scneario where table cannot be deleted because in use; Do not call list_tables on every spec if there are no tables to delete; Stop clearing all table cache when a single table is deleted, manage cache atomically.

+++ spec/dynamodb_local.rb
@@ -34,9 +34,11 @@ module DynamoDBLocal
   end
 
   def self.delete_all_specified_tables!
-    Dynamoid.adapter.list_tables.each do |table|
-      Dynamoid.adapter.delete_table(table) if table =~ /^#{Dynamoid::Config.namespace}/
+    if !Dynamoid.adapter.tables.empty?
+      Dynamoid.adapter.list_tables.each do |table|
+        Dynamoid.adapter.delete_table(table) if table =~ /^#{Dynamoid::Config.namespace}/
+      end
+      Dynamoid.adapter.tables.clear
     end
-    Dynamoid.adapter.tables.clear
   end
 end
